+++ oshi-core/src/main/java/module-info.java
@@ -24,7 +24,6 @@ module com.github.oshi {
     exports oshi.util.platform.mac;
     exports oshi.util.platform.unix.freebsd;
     exports oshi.util.platform.unix.dragonflybsd;
-    exports oshi.util.platform.unix.netbsd;
     exports oshi.util.platform.unix.openbsd;
     exports oshi.util.platform.unix.solaris;
     exports oshi.util.platform.windows;
@@ -44,6 +43,5 @@ module com.github.oshi {
     requires transitive com.sun.jna;
     requires transitive com.sun.jna.platform;
     requires transitive java.desktop;
-    requires java.management;
     requires org.slf4j;
 }